requirements.txt
altgraph==0.17.4
bcrypt==4.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
buildozer==1.5.0
cachetools==5.5.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
Cython==0.29.19
distlib==0.3.8
docutils==0.21.2
dropbox==12.0.2
filelock==3.15.4
filetype==1.2.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
fpdf==1.7.2
fsspec==2025.5.0
google==3.0.0
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
greenlet==3.1.1
gunicorn==23.0.0
httplib2==0.22.0
huggingface-hub==0.31.4
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.5
Kivy==2.3.1
Kivy-Garden==0.1.5
kivy_deps.angle==0.4.0
kivy_deps.glew==0.3.1
kivy_deps.sdl2==0.8.0
KivyGradient==0.0.5
kivymd==1.2.0
MarkupSafe==3.0.2
mercadopago==2.3.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.2.2
packaging==24.1
pdfkit==1.0.0
pefile==2023.2.7
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
ply==3.11
proto-plus==1.24.0
protobuf==5.28.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyparsing==3.1.4
pypiwin32==223
pysnmp==7.1.17
python-dotenv==1.1.0
pytz==2025.2
pywin32==308
pywin32-ctypes==0.2.2
PyYAML==6.0.2
qrcode==8.1
regex==2024.11.6
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.5.3
setuptools==71.1.0
sh==2.0.7
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.38
stone==3.3.1
sympy==1.14.0
timedelta==2020.12.3
tk==0.1.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
vercel==0.2.1
virtualenv==20.26.3
Werkzeug==3.1.3
